refactor: MimeType 엔티티 위치 변경
