+++ packages/lib-mime/src/domain/MimeType.ts
