+++ packages/lib-mime/test/domain/MimeType.test.ts
